Signature helper
Tries to provide the function signature helper when the user is filling parameters.

Given the cursor position, we try to find the first parameter of the list (in clojure, that's the operand) and the position of the parameter the user is filling in. We do that with reverse parsing since, at the cursor position, we don't have a complete clojure expression yet.

With the operand name, we get the documentation with its possible parameters to build the signature helper. With the parameter position, we try to find the most appropriate signature for the user (the signature with the least number of parameters still suitable for the given parameter position).

Closes #8