Add a channel for nREPL output to the Output Window (#92)
Add a channel to show nREPL output to the user. This channel will be
used for the following:

stdout:

Any stdout messages are sent to the channel.

stderr:

Any stderr messages are sent to the channel, and the channel is brought
to the front to make the user aware.

exit:

If the nREPL closes before be we can load the port, then the exit code
is printed to the nREPL output channel, and the channel is brough to the
front.

Fixes #52
Fixes #53
Fixed #69

Also fixes a bug where the 'Starting nREPL' status message would never
clear if the repl failed to start.