Recover from errors more nicely.
There are two changes here.

Firstly, detect when the call to get a stacktrace fails with `unknown-op` - this
indicates that the cider middleware to get stack traces is not loaded.

This lead to the second change: after I fixed the first issue (which is bug #112
I noticed that if you do not have the clojure stack trace middleware, you could
not see the reason for compilation errors. So rather than get a stack trace when
compilation fails, we can print the error message from the compilation operation.

This makes compilation errors take only 1 line on the output panel, which is much
nicer.

Fixes #112